feat: add validation summary to integrated objects (#874)
