+++ app/components/Table/components/TableCell/components/ValidationStatusCell/components/ValidationStatus/entities.ts
@@ -1,0 +1,5 @@
+import { FILE_VALIDATION_STATUS } from "../../../../../../../../apis/catalog/hca-atlas-tracker/common/entities";
+
+export interface Props {
+  validationStatus: FILE_VALIDATION_STATUS;
+}
