+++ app/components/Table/components/TableCell/components/ValidationStatusCell/components/ValidationSummary/entities.ts
@@ -1,0 +1,5 @@
+import { FileValidationSummary } from "../../../../../../../../apis/catalog/hca-atlas-tracker/common/entities";
+
+export interface Props {
+  validationSummary: FileValidationSummary | null;
+}
