+++ app/components/Table/components/TableCell/components/ValidationStatusCell/components/ValidationSummary/validationSummary.styles.ts
