+++ app/components/Table/components/TableCell/components/ValidationStatusCell/components/ValidationSummary/validationSummary.tsx
@@ -1,13 +1,13 @@
 import { SuccessIcon } from "@databiosphere/findable-ui/lib/components/common/CustomIcon/components/SuccessIcon/successIcon";
 import { SVG_ICON_PROPS } from "@databiosphere/findable-ui/lib/styles/common/mui/svgIcon";
 import { Stack } from "@mui/material";
-import { FILE_VALIDATOR_NAME_LABEL } from "../../../../../../apis/catalog/hca-atlas-tracker/common/constants";
-import { FileValidatorName } from "../../../../../../apis/catalog/hca-atlas-tracker/common/entities";
+import { FILE_VALIDATOR_NAME_LABEL } from "../../../../../../../../apis/catalog/hca-atlas-tracker/common/constants";
+import { FileValidatorName } from "../../../../../../../../apis/catalog/hca-atlas-tracker/common/entities";
 import { INNER_STACK_PROPS, STACK_PROPS } from "./constants";
 import { Props } from "./entities";
-import { StyledErrorIcon } from "./validationSummaryCell.styles";
+import { StyledErrorIcon } from "./validationSummary.styles";
 
-export const ValidationSummaryCell = ({
+export const ValidationSummary = ({
   validationSummary,
 }: Props): JSX.Element | null => {
   if (!validationSummary) return null;
