--- app/components/Table/components/TableCell/components/ValidationSummaryCell/entities.ts
@@ -1,5 +1,0 @@
-import { FileValidationSummary } from "../../../../../../apis/catalog/hca-atlas-tracker/common/entities";
-
-export interface Props {
-  validationSummary: FileValidationSummary | null;
-}
