+++ app/views/AtlasSourceDatasetsView/components/Table/columns.ts
@@ -11,10 +11,7 @@ import {
   buildTissue,
 } from "../../../../viewModelBuilders/catalog/hca-atlas-tracker/common/viewModelBuilders";
 import { AtlasSourceDataset } from "../../entities";
-import {
-  renderSourceDatasetCellCount,
-  renderValidationStatus,
-} from "./viewBuilders";
+import { renderSourceDatasetCellCount } from "./viewBuilders";
 
 const COLUMN_ASSAY = {
   accessorKey: "assay",
@@ -111,7 +108,7 @@ const COLUMN_TITLE = {
 
 const COLUMN_VALIDATION_STATUS = {
   accessorKey: "validationStatus",
-  cell: renderValidationStatus,
+  cell: C.ValidationStatusCell,
   header: "Validation Summary",
   meta: { width: { max: "0.5fr", min: "120px" } },
 } as ColumnDef<AtlasSourceDataset>;
